feat: Add support for react 0.14.x
Since the structure of the react tree has changed this approach to
checking if a component has another component as a child is no longer
feasible. This commit introduces a fall back that looks for 'param-case'
of the display name as a class. e.g. AwesomePage would need a class
'awesome-page'. This is mostly introduced to ease the upgrade path for
projects that rely heavily on the have.component check. I would advise
writing new tests in another manner or maybe writing the class check
explicit in your own tests.